+++ src/main/java/com/team4/localer/controller/GroupController.java
@@ -2,8 +2,6 @@ package com.team4.localer.controller;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.apache.ibatis.session.SqlSession;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
@@ -13,10 +11,8 @@ import com.team4.localer.vo.GroupVO;
 @Controller
 public class GroupController {
 
-	@Autowired
-	SqlSession sqlSession;
-	
-	GroupVO vo = new GroupVO();
+//	@Autowired
+//	SqlSession sqlSession;
 	
 	@RequestMapping("/groupPage")
 	public String groupOpen() {
@@ -50,6 +46,27 @@ public class GroupController {
 		mav.setViewName("group/groupWriteForm");
 		return mav;
 	}
+	
+	@RequestMapping("/eatPageView")
+	public ModelAndView eatPageView(int no, String searchKey, String searchWord) {
+		ModelAndView mav = new ModelAndView();
+		
+		
+		mav.setViewName("group/eatPageView");
+		return mav;
+	}
+	
+	
+	
+	@RequestMapping("/withPageView")
+	public ModelAndView withPageView(int no, String searchKey, String searchWord) {
+		ModelAndView mav = new ModelAndView();
+		
+		mav.setViewName("group/withPageView");
+		return mav;
+	}
+	
+	
 	
 	
 }
