--- src/main/java/com/team4/localer/group/GroupVO.java
@@ -1,20 +1,0 @@
-package com.team4.localer.group;
-
-public class GroupVO {
-	private String g_loc;
-	private String up_cate;
-	public String getG_loc() {
-		return g_loc;
-	}
-	public void setG_loc(String g_loc) {
-		this.g_loc = g_loc;
-	}
-	public String getUp_cate() {
-		return up_cate;
-	}
-	public void setUp_cate(String up_cate) {
-		this.up_cate = up_cate;
-	}
-	
-	
-}
